wallet-btc-server/requirements.txt
APScheduler==3.6.0
astroid==2.2.5
certifi==2019.6.16
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
gevent==1.4.0
greenlet==0.4.15
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
pylint==2.3.1
pytz==2019.1
PyYAML==5.1.1
requests==2.22.0
six==1.12.0
typed-ast==1.4.0
tzlocal==1.5.1
urllib3==1.25.3
Werkzeug==0.15.4
wrapt==1.11.2
